+++ test/js/runner.js
@@ -4,7 +4,7 @@ page.onConsoleMessage = function (message) {
     console.log(message);
 };
 
-page.open('test/js/js.html', function (status) {
+page.open('test/js/tests.html', function (status) {
     console.log("\nTesting datascript.js\n");
     var result = page.evaluate(function() {
         return run_tests();