+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# WIP
+# 0.8.0
 
 - Find specifications: collection `:find [?e ...]`, tuple `:find [?e ?v]`, and scalar `:find ?e .`
 - Support for `:db/isComponent` (issue #3)